feat: complete query param implem
